modified according to heller's comment.
