+++ tencentcloud/provider.go
@@ -432,6 +432,7 @@ Monitor
     tencentcloud_monitor_tmp_recording_rule
 	tencentcloud_monitor_tmp_tke_template
 	tencentcloud_monitor_tmp_tke_alert_policy
+	tencentcloud_monitor_tmp_tke_config
 
 PostgreSQL
   Data Source
@@ -1011,7 +1012,6 @@ func Provider() terraform.ResourceProvider {
 			"tencentcloud_kubernetes_scale_worker":                 resourceTencentCloudTkeScaleWorker(),
 			"tencentcloud_kubernetes_cluster_attachment":           resourceTencentCloudTkeClusterAttachment(),
 			"tencentcloud_kubernetes_node_pool":                    ResourceTencentCloudKubernetesNodePool(),
-			"tencentcloud_tke_tmp_config":                          resourceTencentCloudTkeTmpConfig(),
 			"tencentcloud_mysql_backup_policy":                     resourceTencentCloudMysqlBackupPolicy(),
 			"tencentcloud_mysql_account":                           resourceTencentCloudMysqlAccount(),
 			"tencentcloud_mysql_account_privilege":                 resourceTencentCloudMysqlAccountPrivilege(),
@@ -1093,6 +1093,7 @@ func Provider() terraform.ResourceProvider {
 			"tencentcloud_monitor_tmp_recording_rule":              resourceTencentCloudMonitorTmpRecordingRule(),
 			"tencentcloud_monitor_tmp_tke_template":                resourceTencentCloudMonitorTmpTkeTemplate(),
 			"tencentcloud_monitor_tmp_tke_alert_policy":            resourceTencentCloudMonitorTmpTkeAlertPolicy(),
+			"tencentcloud_monitor_tmp_tke_config":                  resourceTencentCloudMonitorTmpTkeConfig(),
 			"tencentcloud_mongodb_standby_instance":                resourceTencentCloudMongodbStandbyInstance(),
 			"tencentcloud_elasticsearch_instance":                  resourceTencentCloudElasticsearchInstance(),
 			"tencentcloud_postgresql_instance":                     resourceTencentCloudPostgresqlInstance(),
