+++ website/tencentcloud.erb
@@ -1146,6 +1146,9 @@
                                     <a href="/docs/providers/tencentcloud/r/monitor_tmp_tke_alert_policy.html">tencentcloud_monitor_tmp_tke_alert_policy</a>
                                 </li>
                                 <li>
+                                    <a href="/docs/providers/tencentcloud/r/monitor_tmp_tke_config.html">tencentcloud_monitor_tmp_tke_config</a>
+                                </li>
+                                <li>
                                     <a href="/docs/providers/tencentcloud/r/monitor_tmp_tke_template.html">tencentcloud_monitor_tmp_tke_template</a>
                                 </li>
                             </ul>
